now edit order is fixed two things : no more modify the created date , second now you cant edit an order thats not pending

+++ app/dashboard/(pages)/orders/edit/[id]/page.tsx
@@ -11,7 +11,7 @@ import axios from "axios";
 import { Alert, AlertDescription, AlertTitle } from "@/components/ui/alert";
 
 import { Order } from "@/types/order";
-import { collection, deleteDoc, getDocs, query, Timestamp, where } from "firebase/firestore";
+import { collection, deleteDoc, getDocs, query, where } from "firebase/firestore";
 import { useOrderStore } from "@/store/orders";
 import ItemsTable from "./components/ItemsTable";
 import { getTotalPriceFromItem } from "@/lib/orders";
@@ -21,24 +21,27 @@ import { useSession } from "next-auth/react";
 import { toast } from "@/hooks/use-toast";
 import { useQuery } from "@tanstack/react-query";
 import { db } from "@/firebase";
+import { StateChanger } from "../../components/StateChanger";
+
+
 
 
 export default function CreateOrder() {
   const { newOrder: order, setNewOrder: setOrder } = useOrderStore();
-  const {id} = useParams<{id: string}>();
+  const { id } = useParams<{ id: string }>();
   const { storeId, store } = useStore();
   const { data: session } = useSession();
 
 
 
-  const {  } = useQuery({
+  const { } = useQuery({
     queryKey: ["order", storeId, id],
     queryFn: async () => {
       if (!storeId) return;
-      if( !id) return;
+      if (!id) return;
       const o = await getDocs(
         query(collection(db, "orders"), where("storeId", "==", storeId),
-        where("sequence", "==", parseInt(id))
+          where("sequence", "==", parseInt(id))
         ),
       ).then(async (response) => {
         const data = response.docs.map((doc) => ({ ...doc.data(), id: doc.id }) as Order);
@@ -81,8 +84,9 @@ export default function CreateOrder() {
 
   const handleSubmit = async () => {
     // get all the inventory items and delete them before creating new ones
-    if(!order) return;
-    const inventoryItems =await getDocs(query(collection(db, "inventoryItems"), where("orderId", "==", order.id)));
+    if (!order) return;
+    if (order.orderStatus !== "pending") return;
+    const inventoryItems = await getDocs(query(collection(db, "inventoryItems"), where("orderId", "==", order.id)));
     inventoryItems.forEach(async (doc) => {
       await deleteDoc(doc.ref);
     });
@@ -127,10 +131,8 @@ export default function CreateOrder() {
     axios
       .post("/api/orders?storeId=" + storeId + "&update=true", {
         ...orderForUpdate,
-        createdAt: order?.createdAt,
-        updatedAt: Timestamp.now(),
-        }
-        )
+      }
+      )
       .then((res) => {
         console.log(res);
       })
@@ -154,6 +156,21 @@ export default function CreateOrder() {
   return (
     order && (
       <div className="space-y-6  mx-auto p-6 bg-white rounded-lg shadow">
+        <div className="flex gap-2 justify-between">
+          <h1 className="text-3xl font-bold">Order #{order.sequence}</h1>
+          <StateChanger showNumberOfCalls state={order.orderStatus} order={order} />
+        </div>
+        {
+          order.orderStatus !== "pending" && (
+            <Alert variant="destructive">
+              <AlertCircle className="h-4 w-4" />
+              <AlertTitle>Warning</AlertTitle>
+              <AlertDescription>
+                You cant modify the order if its not pending
+              </AlertDescription>
+            </Alert>
+          )
+        }
         <div className="flex max-w-7xl gap-4">
           <div className="space-y-4 flex-1 max-w-3xl">
             <h2 className="text-lg font-bold">Customer Information</h2>
@@ -278,7 +295,9 @@ export default function CreateOrder() {
           </Alert>
         )}
 
-        <Button onClick={handleSubmit} className="min-w-xl">
+        <Button
+          disabled={order.orderStatus !== "pending"}
+          onClick={handleSubmit} className="min-w-xl">
           Place Order
         </Button>
       </div>
